backend/requirements.txt
fastapi
uvicorn
networkx
langchain
langchain_community
langchain_ollama
python-multipart
python-dotenv
Pillow
